Project Folder initial commit
If you run it, you'll get a basis of what the window will look like.
There is no functionality apart from being able to drag it around the
screen and increase/decrease the size with the scroll wheel. There is no
x button so you'll have to stop the execution with Visual Studio.