Generator structure + Song Methods/Constructors/More
The overall structure of the generator for this stage of the project is
nearly finalized. However a lot of code is needed to fill in that
structure (plus a few structural things need to be finished). Also the
Song class had a lot of random things added to it to make it actually
usable by the generator.